+++ CHANGELOG.md
@@ -7,3 +7,4 @@ Unreleased
  - System query is now supported when the System registry is present.
  - Service query supports the `--show-service-uri/-u` flag.
  - Setup and installation now set up the docker subnetwork and container ip addresses.
+ - Setup, installation, and running will now default to installing local clouds in `~/.config/pyrrowhead/local-clouds`, but this can be overridden if another install location is desired.